feat: add multiple addresses to identity

+++ services/identity-service/src/controller/identifier/delete-identifier.ts
@@ -2,6 +2,7 @@ import Joi from "joi";
 import { ControllerResponse } from "@lindorm-io/koa";
 import { JOI_GUID } from "../../common";
 import { ServerKoaController } from "../../types";
+import { ClientError } from "@lindorm-io/errors";
 
 interface RequestData {
   id: string;
@@ -20,6 +21,12 @@ export const deleteIdentifierController: ServerKoaController<RequestData> = asyn
     entity: { identifier },
     repository: { identifierRepository },
   } = ctx;
+
+  if (identifier.primary) {
+    throw new ClientError("Invalid request", {
+      description: "Unable to delete primary identifier",
+    });
+  }
 
   await identifierRepository.destroy(identifier);
 };
